requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alphafold-colabfold==2.3.11
appdirs==1.4.4
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
biopython==1.85
biotite==1.2.0
biotraj==1.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
chex==0.1.87
colabfold @ git+https://github.com/sokrypton/ColabFold@ea3292b92ad1648aabccf6d5c96a5057bf76a9d9
contourpy==1.3.2
cycler==0.12.1
dm-haiku==0.0.13
dm-tree==0.1.8
et_xmlfile==2.0.0
fair-esm==2.0.0
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
immutabledict==4.2.2
importlib_metadata==8.7.0
jax==0.4.34
jax-cuda12-pjrt==0.4.34
jax-cuda12-plugin==0.4.34
jax-triton==0.2.0
jaxlib==0.4.34
jaxtyping==0.2.34
Jinja2==3.1.6
jmp==0.0.4
joblib==1.5.2
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
lightning-utilities==0.15.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.0
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
namex==0.1.0
networkx==3.3
numpy==2.1.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.6.77
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
OpenMM==8.2.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.2.3
pdbfixer @ file:///home/conda/feedstock_root/build_artifacts/pdbfixer_1761839329598/work
pillow==11.0.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.2.1
py3Dmol==2.5.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.0
pytz==2025.2
PyYAML==6.0.3
rdkit==2024.3.5
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.14.1
seaborn==0.13.2
silence_tensorflow==1.2.3
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_cpu==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tmtools==0.3.0
tokenizers==0.22.1
toolz==1.0.0
torch==2.9.1
torchaudio==2.1.0+cu118
torchmetrics==1.8.2
torchvision==0.16.0+cu118
tqdm==4.67.0
transformers==4.57.3
triton==3.5.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.1
yarl==1.22.0
zipp==3.23.0
zstandard==0.23.0
